requirements.txt
asgiref==3.5.2
boto3==1.26.54
botocore==1.29.54
click==8.1.3
distlib==0.3.4
Django==4.1.3
django-cors-headers==3.13.0
django-countries==7.4.2
django-environ==0.9.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dotenv-cli==3.1.0
filelock==3.4.2
gunicorn==20.1.0
jmespath==1.0.1
Pillow==9.3.0
platformdirs==2.5.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.13.1
whitenoise==6.3.0
